Splitting out helpers and organizing tests
