+++ crates/agentgateway/src/http/ext_proc.rs
@@ -1,32 +1,24 @@
 use std::convert::Infallible;
-use std::pin::Pin;
-use std::task::{Context, Poll};
 
 use anyhow::anyhow;
-use bytes::{Buf, Bytes};
-use http_body::{Body, Frame, SizeHint};
+use bytes::Bytes;
+use http_body::{Body, Frame};
 use http_body_util::BodyExt;
-use itertools::Itertools;
-use pin_project_lite::pin_project;
 use prost_wkt_types::Struct;
-use proto::body_mutation::Mutation;
 use proto::processing_request::Request;
 use proto::processing_response::Response;
-use protos::envoy::service::ext_proc::v3::{
-	BodySendMode as EnvoyBodySendMode, ProtocolConfiguration,
-};
+use protos::envoy::service::ext_proc::v3::ProtocolConfiguration;
 use serde_json::Value as JsonValue;
 use tokio::sync::mpsc::{Receiver, Sender};
 use tokio_stream::wrappers::ReceiverStream;
 
 use crate::cel::{Executor, Expression, RequestSnapshot};
 use crate::client::ResolvedDestination;
-use crate::http::buflist::BufList;
 use crate::http::ext_proc::proto::{
-	BodyMutation, BodyResponse, HeaderMutation, HeadersResponse, HttpBody, HttpHeaders, HttpTrailers,
-	ImmediateResponse, Metadata, ProcessingRequest, ProcessingResponse, processing_response,
+	HttpBody, HttpHeaders, HttpTrailers, Metadata, ProcessingRequest, ProcessingResponse,
+	processing_response,
 };
-use crate::http::{HeaderName, PolicyResponse, bufferbody, envoy_proto_common};
+use crate::http::{HeaderName, PolicyResponse, envoy_proto_common};
 use crate::proxy::ProxyError;
 use crate::proxy::dtrace::{self, pol_result};
 use crate::proxy::httpproxy::PolicyClient;
@@ -52,6 +44,8 @@ pub enum Error {
 	ResponseDropped,
 	#[error("failed to buffer body: {0}")]
 	BodyBuffer(String),
+	#[error("invalid body mutation: {0}")]
+	BodyMutation(String),
 	#[error("failed to convert metadata value: {0}")]
 	MetadataConversion(String),
 	#[error(transparent)]
@@ -73,79 +67,34 @@ pub mod proto {
 	pub use protos::envoy::service::ext_proc::v3::*;
 }
 
-#[apply(schema!)]
-#[derive(Default, Copy, PartialEq, Eq)]
-pub enum FailureMode {
-	#[default]
-	FailClosed,
-	FailOpen,
-}
+mod buffering;
+mod channel;
+mod headers;
+mod mutation;
+mod processing;
 
-#[apply(schema!)]
-#[derive(Default, Copy, PartialEq, Eq)]
-pub enum BodySendMode {
-	None,
-	Buffered,
-	BufferedPartial,
-	#[default]
-	FullDuplexStreamed,
-}
-
-impl From<BodySendMode> for EnvoyBodySendMode {
-	fn from(val: BodySendMode) -> Self {
-		match val {
-			BodySendMode::None => EnvoyBodySendMode::None,
-			BodySendMode::Buffered => EnvoyBodySendMode::Buffered,
-			BodySendMode::BufferedPartial => EnvoyBodySendMode::BufferedPartial,
-			BodySendMode::FullDuplexStreamed => EnvoyBodySendMode::FullDuplexStreamed,
-		}
-	}
-}
-
-impl From<BodySendMode> for i32 {
-	fn from(val: BodySendMode) -> Self {
-		match val {
-			BodySendMode::None => EnvoyBodySendMode::None as i32,
-			BodySendMode::Buffered => EnvoyBodySendMode::Buffered as i32,
-			BodySendMode::BufferedPartial => EnvoyBodySendMode::BufferedPartial as i32,
-			BodySendMode::FullDuplexStreamed => EnvoyBodySendMode::FullDuplexStreamed as i32,
-		}
-	}
-}
-
-impl From<EnvoyBodySendMode> for BodySendMode {
-	fn from(val: EnvoyBodySendMode) -> Self {
-		match val {
-			EnvoyBodySendMode::None => BodySendMode::None,
-			EnvoyBodySendMode::Buffered => BodySendMode::Buffered,
-			EnvoyBodySendMode::BufferedPartial => BodySendMode::BufferedPartial,
-			EnvoyBodySendMode::FullDuplexStreamed => BodySendMode::FullDuplexStreamed,
-			EnvoyBodySendMode::Streamed => {
-				// This mode is not currently supported, so we map it to the closest available mode and log a warning.
-				warn!(
-					"received unsupported Envoy body send mode STREAMED; treating as FULL_DUPLEX_STREAMED"
-				);
-				BodySendMode::FullDuplexStreamed
-			},
-		}
-	}
-}
-
-#[apply(schema!)]
-#[derive(Default, Copy, PartialEq, Eq)]
-pub enum HeaderSendMode {
-	#[default]
-	Send,
-	Skip,
-}
-
-#[apply(schema!)]
-#[derive(Default, Copy, PartialEq, Eq)]
-pub enum TrailerSendMode {
-	Send,
-	#[default]
-	Skip,
-}
+use buffering::{
+	BufferedBodyPhase, PendingBufferedBody, attach_request_body_channel,
+	debug_assert_preserved_request_body, start_buffered_request_body, start_buffered_response_body,
+};
+pub use channel::GrpcReferenceChannel;
+use headers::{req_to_header_map, resp_to_header_map, to_header_map};
+#[cfg(test)]
+use mutation::extract_dynamic_metadata;
+use mutation::{
+	handle_response_for_request_mutation, handle_response_for_response_mutation,
+	request_body_response_has_no_mutation, request_response_has_streamed_body_mutation,
+	response_body_response_has_no_mutation, response_response_has_streamed_body_mutation,
+	to_immediate_response,
+};
+use processing::{
+	BodyPath, BodyStreamDirection, FirstExtProcMessage, HeaderPhase, ModeStateMachine,
+	RequestFlowFsm, RequestLoopStep, RequestPhase, ResponseFlowFsm, ResponseLoopMessage,
+	ResponsePhase,
+};
+pub use processing::{
+	BodySendMode, FailureMode, HeaderSendMode, ProcessingOptions, TrailerSendMode,
+};
 
 #[apply(schema!)]
 #[derive(Default, Copy, PartialEq, Eq)]
@@ -229,6 +178,8 @@ impl InferenceRouting {
 				ProcessingOptions {
 					request_body_mode: BodySendMode::FullDuplexStreamed,
 					response_body_mode: BodySendMode::FullDuplexStreamed,
+					request_trailer_mode: TrailerSendMode::Send,
+					response_trailer_mode: TrailerSendMode::Send,
 					..Default::default()
 				},
 			)),
@@ -276,829 +227,6 @@ impl InferencePoolRouter {
 		*resp = new_resp;
 		Ok(pr.unwrap_or_default())
 	}
-}
-
-#[apply(schema!)]
-#[derive(Copy)]
-pub struct ProcessingOptions {
-	#[serde(default = "default_body_send_mode")]
-	pub request_body_mode: BodySendMode,
-	#[serde(default = "default_body_send_mode")]
-	pub response_body_mode: BodySendMode,
-	#[serde(default)]
-	pub request_header_mode: HeaderSendMode,
-	#[serde(default)]
-	pub response_header_mode: HeaderSendMode,
-	#[serde(default)]
-	pub request_trailer_mode: TrailerSendMode,
-	#[serde(default)]
-	pub response_trailer_mode: TrailerSendMode,
-	/// Allow ext_proc `mode_override` values from matching headers responses to update
-	/// subsequent request/response processing phases for this exchange.
-	#[serde(default)]
-	pub allow_mode_override: bool,
-}
-
-fn default_body_send_mode() -> BodySendMode {
-	BodySendMode::FullDuplexStreamed
-}
-
-impl Default for ProcessingOptions {
-	fn default() -> Self {
-		Self {
-			request_body_mode: BodySendMode::FullDuplexStreamed,
-			response_body_mode: BodySendMode::FullDuplexStreamed,
-			request_header_mode: HeaderSendMode::Send,
-			response_header_mode: HeaderSendMode::Send,
-			request_trailer_mode: TrailerSendMode::Skip,
-			response_trailer_mode: TrailerSendMode::Skip,
-			allow_mode_override: false,
-		}
-	}
-}
-
-#[derive(Debug, Copy, Clone)]
-enum HeaderPhase {
-	Request,
-	Response,
-}
-
-#[derive(Debug, Copy, Clone)]
-enum BodyModeOverrideDirection {
-	Request,
-	Response,
-}
-
-// Tracks the negotiated ext_proc processing modes for the lifetime of a single
-// gRPC bidirectional stream. This is the protocol-facing state: it stores what we believe
-// Envoy ext_proc wants us to send next, and applies mode_override updates with
-// protocol validity checks.
-#[derive(Debug, Copy, Clone)]
-struct ModeStateMachine {
-	request_body_mode: BodySendMode,
-	response_body_mode: BodySendMode,
-	request_header_mode: HeaderSendMode,
-	response_header_mode: HeaderSendMode,
-	request_trailer_mode: TrailerSendMode,
-	response_trailer_mode: TrailerSendMode,
-	allow_mode_override: bool,
-	request_headers_processed: bool,
-	response_headers_processed: bool,
-}
-
-impl From<ProcessingOptions> for ModeStateMachine {
-	fn from(opts: ProcessingOptions) -> Self {
-		Self {
-			request_body_mode: opts.request_body_mode,
-			response_body_mode: opts.response_body_mode,
-			request_header_mode: opts.request_header_mode,
-			response_header_mode: opts.response_header_mode,
-			request_trailer_mode: opts.request_trailer_mode,
-			response_trailer_mode: opts.response_trailer_mode,
-			allow_mode_override: opts.allow_mode_override,
-			request_headers_processed: false,
-			response_headers_processed: false,
-		}
-	}
-}
-
-impl ModeStateMachine {
-	fn current_mode_allows_body_override(
-		current_mode: BodySendMode,
-		phase: HeaderPhase,
-		direction: BodyModeOverrideDirection,
-	) -> bool {
-		if current_mode == BodySendMode::FullDuplexStreamed {
-			warn!(
-				phase = ?phase,
-				direction = ?direction,
-				"ignoring body mode_override because current body mode is FULL_DUPLEX_STREAMED"
-			);
-			return false;
-		}
-		true
-	}
-
-	fn apply_body_mode_override(
-		current_mode: &mut BodySendMode,
-		next_mode: EnvoyBodySendMode,
-		phase: HeaderPhase,
-		direction: BodyModeOverrideDirection,
-		phase_name: &'static str,
-		field_name: &'static str,
-	) {
-		match next_mode {
-			// TODO: Should we abort the stream?
-			EnvoyBodySendMode::Streamed => {
-				warn!(
-					phase = phase_name,
-					field = field_name,
-					"mode_override body_mode=STREAMED is not implemented; keeping current body mode"
-				)
-			},
-			_ if Self::current_mode_allows_body_override(*current_mode, phase, direction) => {
-				*current_mode = BodySendMode::from(next_mode);
-			},
-			_ => {},
-		}
-	}
-
-	fn mark_headers_processed(&mut self, phase: HeaderPhase) {
-		match phase {
-			HeaderPhase::Request => self.request_headers_processed = true,
-			HeaderPhase::Response => self.response_headers_processed = true,
-		}
-	}
-
-	fn apply_envoy_mode_override(&mut self, phase: HeaderPhase, mode: &proto::ProcessingMode) {
-		use proto::processing_mode::HeaderSendMode as EnvoyHeaderSendMode;
-		let phase_name = match phase {
-			HeaderPhase::Request => "request_headers",
-			HeaderPhase::Response => "response_headers",
-		};
-
-		if mode.request_header_mode != EnvoyHeaderSendMode::Default as i32 {
-			warn!(
-				phase = phase_name,
-				"mode_override.request_header_mode is ignored by protocol"
-			);
-		}
-
-		if let Ok(hm) = EnvoyHeaderSendMode::try_from(mode.response_header_mode) {
-			match hm {
-				EnvoyHeaderSendMode::Default => {},
-				EnvoyHeaderSendMode::Send if !self.response_headers_processed => {
-					self.response_header_mode = HeaderSendMode::Send;
-				},
-				EnvoyHeaderSendMode::Skip if !self.response_headers_processed => {
-					self.response_header_mode = HeaderSendMode::Skip;
-				},
-				_ => {},
-			}
-		}
-
-		if let Ok(bm) = EnvoyBodySendMode::try_from(mode.request_body_mode) {
-			Self::apply_body_mode_override(
-				&mut self.request_body_mode,
-				bm,
-				phase,
-				BodyModeOverrideDirection::Request,
-				phase_name,
-				"request_body_mode",
-			);
-		}
-
-		if let Ok(bm) = EnvoyBodySendMode::try_from(mode.response_body_mode) {
-			Self::apply_body_mode_override(
-				&mut self.response_body_mode,
-				bm,
-				phase,
-				BodyModeOverrideDirection::Response,
-				phase_name,
-				"response_body_mode",
-			);
-		}
-
-		if let Ok(hm) = EnvoyHeaderSendMode::try_from(mode.request_trailer_mode) {
-			match hm {
-				EnvoyHeaderSendMode::Default => {},
-				EnvoyHeaderSendMode::Send => self.request_trailer_mode = TrailerSendMode::Send,
-				EnvoyHeaderSendMode::Skip => self.request_trailer_mode = TrailerSendMode::Skip,
-			}
-		}
-		if let Ok(hm) = EnvoyHeaderSendMode::try_from(mode.response_trailer_mode) {
-			match hm {
-				EnvoyHeaderSendMode::Default => {},
-				EnvoyHeaderSendMode::Send => self.response_trailer_mode = TrailerSendMode::Send,
-				EnvoyHeaderSendMode::Skip => self.response_trailer_mode = TrailerSendMode::Skip,
-			}
-		}
-	}
-}
-
-// Request-side execution FSM. Unlike ModeStateMachine (which stores ext_proc
-// protocol configuration), this FSM captures local control-flow transitions in
-// mutate_request: whether we are waiting on headers, body, or can return.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-enum RequestPhase {
-	AwaitingHeaders,
-	AwaitingBody,
-	StreamingContinuation,
-	Complete,
-}
-
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-enum BodyPath {
-	None,
-	Buffered,
-	BufferedPartial,
-	FullDuplex,
-}
-
-#[derive(Debug, Copy, Clone)]
-struct RequestFlowFsm {
-	phase: RequestPhase,
-	body_path: BodyPath,
-	had_body: bool,
-	expect_body_response: bool,
-}
-
-impl RequestFlowFsm {
-	fn new(send_headers: bool, had_body: bool, body_mode: BodySendMode) -> Self {
-		let body_path = BodyPath::from(body_mode);
-		let expect_body_response = body_path != BodyPath::None && had_body;
-		let phase = if send_headers {
-			RequestPhase::AwaitingHeaders
-		} else if expect_body_response {
-			RequestPhase::AwaitingBody
-		} else {
-			RequestPhase::Complete
-		};
-		Self {
-			phase,
-			body_path,
-			had_body,
-			expect_body_response,
-		}
-	}
-
-	fn reconcile_potential_mode_override(&mut self, mode: BodySendMode) {
-		self.body_path = BodyPath::from(mode);
-		self.expect_body_response = self.body_path != BodyPath::None && self.had_body;
-		if self.phase == RequestPhase::AwaitingBody && !self.expect_body_response {
-			self.phase = RequestPhase::Complete;
-		}
-	}
-
-	fn finish_headers_phase(&mut self) {
-		self.phase = if self.expect_body_response {
-			RequestPhase::AwaitingBody
-		} else {
-			RequestPhase::Complete
-		};
-	}
-
-	fn enter_streaming_continuation(&mut self) {
-		self.phase = RequestPhase::StreamingContinuation;
-	}
-
-	fn advance_after_response(&mut self, headers_done: bool) -> bool {
-		if headers_done {
-			self.finish_headers_phase();
-		}
-		self.phase != RequestPhase::AwaitingHeaders
-	}
-
-	fn should_restore_original_buffered_body(&self, body_no_mutation: bool) -> bool {
-		body_no_mutation && self.expect_body_response && self.body_path != BodyPath::FullDuplex
-	}
-
-	fn should_fail_open_on_disconnect(&self, failure_mode: FailureMode) -> bool {
-		failure_mode == FailureMode::FailOpen
-			&& (!self.had_body || self.body_path != BodyPath::FullDuplex)
-	}
-}
-
-impl From<BodySendMode> for BodyPath {
-	fn from(mode: BodySendMode) -> Self {
-		match mode {
-			BodySendMode::None => BodyPath::None,
-			BodySendMode::Buffered => BodyPath::Buffered,
-			BodySendMode::BufferedPartial => BodyPath::BufferedPartial,
-			BodySendMode::FullDuplexStreamed => BodyPath::FullDuplex,
-		}
-	}
-}
-
-// Response-side execution FSM mirroring RequestFlowFsm. This controls local
-// phase progression in mutate_response independently of protocol mode storage.
-#[derive(Debug, Copy, Clone, PartialEq, Eq)]
-enum ResponsePhase {
-	AwaitingHeaders,
-	AwaitingBody,
-	StreamingContinuation,
-	Complete,
-}
-
-#[derive(Debug, Copy, Clone)]
-struct ResponseFlowFsm {
-	phase: ResponsePhase,
-	body_path: BodyPath,
-	had_body: bool,
-	send_body: bool,
-}
-
-enum ResponseLoopMessage {
-	Immediate(PolicyResponse),
-	Processing(ProcessingResponse),
-}
-
-struct RequestLoopStep {
-	transitioned: bool,
-	eos: bool,
-	body_no_mutation: bool,
-}
-
-#[derive(Copy, Clone)]
-struct BodyMessageFns {
-	body: fn(HttpBody) -> Request,
-	trailers: fn(HttpTrailers) -> Request,
-}
-
-impl BodyMessageFns {
-	const REQUEST: Self = Self {
-		body: Request::RequestBody,
-		trailers: Request::RequestTrailers,
-	};
-	const RESPONSE: Self = Self {
-		body: Request::ResponseBody,
-		trailers: Request::ResponseTrailers,
-	};
-}
-
-impl ResponseFlowFsm {
-	fn new(send_headers: bool, body_mode: BodySendMode, had_body: bool) -> Self {
-		let body_path = BodyPath::from(body_mode);
-		let send_body = body_path != BodyPath::None && had_body;
-		let phase = if send_headers {
-			ResponsePhase::AwaitingHeaders
-		} else if send_body {
-			ResponsePhase::AwaitingBody
-		} else {
-			ResponsePhase::Complete
-		};
-		Self {
-			phase,
-			body_path,
-			had_body,
-			send_body,
-		}
-	}
-
-	fn reconcile_potential_mode_override(&mut self, mode: BodySendMode) {
-		self.body_path = BodyPath::from(mode);
-		self.send_body = self.body_path != BodyPath::None && self.had_body;
-		if self.phase == ResponsePhase::AwaitingBody && !self.send_body {
-			self.phase = ResponsePhase::Complete;
-		}
-	}
-
-	fn finish_headers_phase(&mut self) {
-		self.phase = if self.send_body {
-			ResponsePhase::AwaitingBody
-		} else {
-			ResponsePhase::Complete
-		};
-	}
-
-	fn enter_streaming_continuation(&mut self) {
-		self.phase = ResponsePhase::StreamingContinuation;
-	}
-
-	fn advance_after_response(&mut self, headers_done: bool) -> bool {
-		if headers_done {
-			self.finish_headers_phase();
-		}
-		self.phase != ResponsePhase::AwaitingHeaders
-	}
-
-	fn should_restore_original_buffered_body(&self, body_no_mutation: bool) -> bool {
-		body_no_mutation && self.send_body && self.body_path != BodyPath::FullDuplex
-	}
-}
-
-#[derive(Debug, Default)]
-struct FirstExtProcMessage {
-	attributes: Option<HashMap<String, Struct>>,
-	protocol_config: Option<ProtocolConfiguration>,
-}
-
-impl FirstExtProcMessage {
-	fn for_body_phase(
-		send_headers: bool,
-		send_body: bool,
-		attributes: HashMap<String, Struct>,
-		protocol_config: ProtocolConfiguration,
-		protocol_config_sent: bool,
-	) -> Self {
-		Self {
-			attributes: (!send_headers && send_body).then_some(attributes),
-			protocol_config: (!send_headers && send_body && !protocol_config_sent)
-				.then_some(protocol_config),
-		}
-	}
-
-	fn take_attributes_or_default(&mut self) -> HashMap<String, Struct> {
-		self.attributes.take().unwrap_or_default()
-	}
-
-	fn take_protocol_config(&mut self) -> Option<ProtocolConfiguration> {
-		self.protocol_config.take()
-	}
-
-	fn has_protocol_config(&self) -> bool {
-		self.protocol_config.is_some()
-	}
-
-	fn take_for_send(message: &mut Self) -> (Self, bool) {
-		let message = std::mem::take(message);
-		let sends_protocol_config = message.has_protocol_config();
-		(message, sends_protocol_config)
-	}
-}
-
-enum BufferedBodyPhase {
-	Deferred {
-		body: http::Body,
-		mode: BodySendMode,
-		limit: usize,
-		error_message: &'static str,
-	},
-	Replaying {
-		original: bytes::Bytes,
-	},
-	PartialReplaying {
-		original: Option<bytes::Bytes>,
-		remainder: Option<http::Body>,
-	},
-}
-
-enum PendingBufferedBody {
-	Body {
-		body: http::Body,
-		handle: bufferbody::BufferedBodyHandle,
-		error_message: &'static str,
-	},
-	Partial {
-		body: bytes::Bytes,
-		end_stream: bool,
-		trailers: Option<::http::HeaderMap>,
-	},
-}
-
-impl BufferedBodyPhase {
-	fn new(body: http::Body, mode: BodySendMode, limit: usize, error_message: &'static str) -> Self {
-		Self::Deferred {
-			body,
-			mode,
-			limit,
-			error_message,
-		}
-	}
-
-	async fn take_pending_send(
-		phase: &mut Option<Self>,
-	) -> Result<Option<PendingBufferedBody>, Error> {
-		let Some(buffered_body) = phase.take() else {
-			return Ok(None);
-		};
-		match buffered_body {
-			Self::Deferred {
-				body,
-				mode,
-				limit,
-				error_message,
-			} => match mode {
-				BodySendMode::Buffered => {
-					let (buffered_body, handle) = bufferbody::BufferedBody::new_with_limit(body, limit);
-					Ok(Some(PendingBufferedBody::Body {
-						body: http::Body::new(buffered_body),
-						handle,
-						error_message,
-					}))
-				},
-				BodySendMode::BufferedPartial => {
-					let buffered = buffer_body_partial(body, limit, error_message).await?;
-					*phase = Some(Self::PartialReplaying {
-						original: Some(buffered.body.clone()),
-						remainder: buffered.remainder,
-					});
-					Ok(Some(PendingBufferedBody::Partial {
-						body: buffered.body,
-						end_stream: buffered.end_stream,
-						trailers: buffered.trailers,
-					}))
-				},
-				_ => Ok(None),
-			},
-			replaying @ Self::Replaying { .. } => {
-				*phase = Some(replaying);
-				Ok(None)
-			},
-			partial @ Self::PartialReplaying { .. } => {
-				*phase = Some(partial);
-				Ok(None)
-			},
-		}
-	}
-
-	fn take_deferred_body(phase: &mut Option<Self>) -> Option<http::Body> {
-		let deferred = phase.take()?;
-		match deferred {
-			Self::Deferred { body, .. } => Some(body),
-			other => {
-				*phase = Some(other);
-				None
-			},
-		}
-	}
-
-	fn update_deferred_mode(phase: &mut Option<Self>, next_mode: BodySendMode) {
-		if let Some(Self::Deferred { mode, .. }) = phase.as_mut() {
-			*mode = next_mode;
-		}
-	}
-
-	fn take_partial_remainder(phase: &mut Option<Self>) -> Option<http::Body> {
-		let buffered = phase.take()?;
-		match buffered {
-			Self::PartialReplaying { remainder, .. } => remainder,
-			other => {
-				*phase = Some(other);
-				None
-			},
-		}
-	}
-
-	fn replay_from_handle(
-		phase: &mut Option<Self>,
-		handle: bufferbody::BufferedBodyHandle,
-	) -> Result<(), Error> {
-		let Some(original) = handle.bytes() else {
-			return Err(Error::BodyBuffer(
-				"buffered body completed without captured bytes".into(),
-			));
-		};
-		*phase = Some(Self::Replaying { original });
-		Ok(())
-	}
-
-	async fn restore_original_if(
-		phase: &mut Option<Self>,
-		restore: bool,
-		body_tx: &mut Sender<Result<Frame<Bytes>, Infallible>>,
-	) {
-		if !restore {
-			return;
-		}
-		let Some(buffered_body) = phase.take() else {
-			return;
-		};
-		match buffered_body {
-			Self::Replaying { original } => {
-				let _ = body_tx.send(Ok(Frame::data(original))).await;
-			},
-			Self::PartialReplaying {
-				mut original,
-				remainder,
-			} => {
-				if let Some(original) = original.take()
-					&& !original.is_empty()
-				{
-					let _ = body_tx.send(Ok(Frame::data(original))).await;
-				}
-				*phase = Some(Self::PartialReplaying {
-					original,
-					remainder,
-				});
-			},
-			other => {
-				*phase = Some(other);
-			},
-		}
-	}
-}
-
-struct BufferedPartialBody {
-	body: bytes::Bytes,
-	end_stream: bool,
-	trailers: Option<::http::HeaderMap>,
-	remainder: Option<http::Body>,
-}
-
-pin_project! {
-	// Body wrapper used when the configured partial-buffer limit falls in the middle of a data
-	// frame. The bytes after the limit have already been read, so replay them before polling the
-	// original body for the rest of the stream.
-	struct RemainderBody {
-		prefix: Option<bytes::Bytes>,
-		#[pin]
-		inner: http::Body,
-	}
-}
-
-impl RemainderBody {
-	fn new(prefix: Option<bytes::Bytes>, inner: http::Body) -> Self {
-		Self { prefix, inner }
-	}
-}
-
-impl Body for RemainderBody {
-	type Data = Bytes;
-	type Error = axum_core::Error;
-
-	fn poll_frame(
-		self: Pin<&mut Self>,
-		cx: &mut Context<'_>,
-	) -> Poll<Option<Result<Frame<Self::Data>, Self::Error>>> {
-		let this = self.project();
-		if let Some(prefix) = this.prefix.take()
-			&& !prefix.is_empty()
-		{
-			return Poll::Ready(Some(Ok(Frame::data(prefix))));
-		}
-		this.inner.poll_frame(cx)
-	}
-
-	fn is_end_stream(&self) -> bool {
-		self.prefix.as_ref().map(Bytes::is_empty).unwrap_or(true) && self.inner.is_end_stream()
-	}
-
-	fn size_hint(&self) -> SizeHint {
-		let prefix_len = self.prefix.as_ref().map(Bytes::len).unwrap_or_default() as u64;
-		let mut hint = self.inner.size_hint();
-		hint.set_lower(hint.lower().saturating_add(prefix_len));
-		if let Some(upper) = hint.upper() {
-			hint.set_upper(upper.saturating_add(prefix_len));
-		}
-		hint
-	}
-}
-
-async fn buffer_body_partial(
-	mut body: http::Body,
-	limit: usize,
-	error_message: &'static str,
-) -> Result<BufferedPartialBody, Error> {
-	// A zero-byte limit still means BufferedPartial should send a body message to ext_proc; it
-	// just contains no data and the original body remains entirely local pass-through.
-	if limit == 0 {
-		return Ok(BufferedPartialBody {
-			body: bytes::Bytes::new(),
-			end_stream: false,
-			trailers: None,
-			remainder: Some(body),
-		});
-	}
-
-	let mut buffer = BufList::default();
-	let mut buffered = 0usize;
-	let mut trailers = None;
-	loop {
-		let Some(frame) = body
-			.frame()
-			.await
-			.transpose()
-			.map_err(|e| Error::BodyBuffer(format!("{error_message}: {e}")))?
-		else {
-			let body = buffered_bytes(buffer);
-			return Ok(BufferedPartialBody {
-				body,
-				end_stream: trailers.is_none(),
-				trailers,
-				remainder: None,
-			});
-		};
-
-		match frame.into_data().map_err(Frame::into_trailers) {
-			Ok(mut data) => {
-				let len = data.remaining();
-				let remaining_limit = limit.saturating_sub(buffered);
-				if len <= remaining_limit {
-					// The whole frame fits in the partial buffer. If this reaches the limit and
-					// the body is not at EOF, stop reading and leave the rest for normal upstream
-					// or downstream forwarding.
-					let bytes = data.copy_to_bytes(len);
-					if bytes.has_remaining() {
-						buffer.push(bytes);
-						buffered += len;
-					}
-					if buffered == limit && !body.is_end_stream() {
-						return Ok(BufferedPartialBody {
-							body: buffered_bytes(buffer),
-							end_stream: false,
-							trailers: None,
-							remainder: Some(body),
-						});
-					}
-				} else {
-					// The frame crosses the partial-buffer boundary. Split it so ext_proc sees
-					// only the prefix, then replay the remaining bytes before polling the body
-					// again.
-					let prefix = data.copy_to_bytes(remaining_limit);
-					if prefix.has_remaining() {
-						buffer.push(prefix);
-					}
-					let rest = data.copy_to_bytes(data.remaining());
-					let remainder = http::Body::new(RemainderBody::new(Some(rest), body));
-					return Ok(BufferedPartialBody {
-						body: buffered_bytes(buffer),
-						end_stream: false,
-						trailers: None,
-						remainder: Some(remainder),
-					});
-				}
-			},
-			Err(Ok(frame_trailers)) => {
-				trailers = Some(frame_trailers);
-			},
-			Err(Err(_unknown)) => {
-				tracing::warn!("An unknown body frame has been buffered");
-				return Ok(BufferedPartialBody {
-					body: buffered_bytes(buffer),
-					end_stream: true,
-					trailers,
-					remainder: None,
-				});
-			},
-		}
-	}
-}
-
-fn buffered_bytes(mut buffer: BufList) -> bytes::Bytes {
-	let len = buffer.remaining();
-	buffer.copy_to_bytes(len)
-}
-
-fn attach_request_body_channel(
-	req: http::Request,
-	rx_chunk: &mut Option<Receiver<Result<Frame<Bytes>, Infallible>>>,
-) -> (http::Request, http::Body) {
-	let (parts, body) = req.into_parts();
-	let rx_chunk = rx_chunk
-		.take()
-		.expect("request body channel should only be attached once");
-	let upstream_body = http_body_util::StreamBody::new(ReceiverStream::new(rx_chunk));
-	(
-		http::Request::from_parts(parts, http::Body::new(upstream_body)),
-		body,
-	)
-}
-
-#[cfg(debug_assertions)]
-fn debug_assert_preserved_request_body(req: &http::Request, had_body: bool, context: &'static str) {
-	if had_body {
-		debug_assert!(
-			!req.body().is_end_stream(),
-			"{context}: returning request should preserve the original non-empty body"
-		);
-	}
-}
-
-#[cfg(not(debug_assertions))]
-fn debug_assert_preserved_request_body(
-	_req: &http::Request,
-	_had_body: bool,
-	_context: &'static str,
-) {
-}
-
-fn should_buffer_body(body_mode: BodySendMode, had_body: bool) -> bool {
-	had_body
-		&& matches!(
-			body_mode,
-			BodySendMode::Buffered | BodySendMode::BufferedPartial
-		)
-}
-
-fn start_buffered_request_body(
-	req: http::Request,
-	body_mode: BodySendMode,
-	had_body: bool,
-	rx_chunk: &mut Option<Receiver<Result<Frame<Bytes>, Infallible>>>,
-) -> (http::Request, Option<BufferedBodyPhase>) {
-	if !should_buffer_body(body_mode, had_body) {
-		return (req, None);
-	}
-	let max_request_bytes = http::buffer_limit(&req);
-	let (req, req_body) = attach_request_body_channel(req, rx_chunk);
-	let phase = BufferedBodyPhase::new(
-		req_body,
-		body_mode,
-		max_request_bytes,
-		"failed to read request body for buffering",
-	);
-	(req, Some(phase))
-}
-
-fn start_buffered_response_body(
-	body: &mut Option<http::Body>,
-	body_mode: BodySendMode,
-	had_body: bool,
-	max_response_bytes: usize,
-) -> Option<BufferedBodyPhase> {
-	if !should_buffer_body(body_mode, had_body) {
-		return None;
-	}
-	let body = body
-		.take()
-		.expect("response body should be available before buffering starts");
-	Some(BufferedBodyPhase::new(
-		body,
-		body_mode,
-		max_response_bytes,
-		"failed to read response body for buffering",
-	))
 }
 
 #[apply(schema!)]
@@ -1342,7 +470,7 @@ impl ExtProcInstance {
 		buffered_body: &mut Option<BufferedBodyPhase>,
 		metadata_context: &Option<Arc<Metadata>>,
 		first_message: &mut FirstExtProcMessage,
-		message_fns: BodyMessageFns,
+		body_direction: BodyStreamDirection,
 		send_trailers: bool,
 	) -> Result<bool, Error> {
 		let Some(pending) = BufferedBodyPhase::take_pending_send(buffered_body).await? else {
@@ -1360,8 +488,7 @@ impl ExtProcInstance {
 					metadata_context.clone(),
 					body,
 					self.tx_req.clone(),
-					message_fns.body,
-					message_fns.trailers,
+					body_direction,
 					send_trailers,
 					first_message,
 					error_message,
@@ -1380,8 +507,7 @@ impl ExtProcInstance {
 				Self::send_partial_body(
 					metadata_context.clone(),
 					self.tx_req.clone(),
-					message_fns.body,
-					message_fns.trailers,
+					body_direction,
 					body,
 					end_stream,
 					trailers,
@@ -1399,8 +525,7 @@ impl ExtProcInstance {
 	async fn send_partial_body(
 		metadata_context: Option<Arc<Metadata>>,
 		tx: Sender<ProcessingRequest>,
-		body_fn: fn(HttpBody) -> Request,
-		trail_fn: fn(HttpTrailers) -> Request,
+		body_direction: BodyStreamDirection,
 		body: bytes::Bytes,
 		end_stream: bool,
 		trailers: Option<::http::HeaderMap>,
@@ -1409,8 +534,8 @@ impl ExtProcInstance {
 	) -> Result<(), Error> {
 		let mut first_message = first_message;
 		tx.send(ProcessingRequest {
-			request: Some(body_fn(HttpBody {
-				body: body.into(),
+			request: Some(body_direction.body_message(HttpBody {
+				body,
 				end_of_stream: end_stream,
 			})),
 			metadata_context: metadata_context.as_deref().cloned(),
@@ -1426,7 +551,7 @@ impl ExtProcInstance {
 		};
 		if send_trailers {
 			tx.send(ProcessingRequest {
-				request: Some(trail_fn(HttpTrailers {
+				request: Some(body_direction.trailers_message(HttpTrailers {
 					trailers: to_header_map(&trailers),
 				})),
 				metadata_context: metadata_context.as_deref().cloned(),
@@ -1436,9 +561,10 @@ impl ExtProcInstance {
 			})
 			.await
 			.map_err(|_| Error::RequestSend)?;
+			return Ok(());
 		}
 		tx.send(ProcessingRequest {
-			request: Some(body_fn(HttpBody {
+			request: Some(body_direction.body_message(HttpBody {
 				body: Default::default(),
 				end_of_stream: true,
 			})),
@@ -1452,12 +578,13 @@ impl ExtProcInstance {
 		Ok(())
 	}
 
+	#[allow(clippy::too_many_arguments)]
 	fn spawn_body_stream(
 		&self,
 		metadata_context: &Option<Arc<Metadata>>,
 		pending_body: &mut Option<http::Body>,
 		tx: Sender<ProcessingRequest>,
-		message_fns: BodyMessageFns,
+		body_direction: BodyStreamDirection,
 		send_trailers: bool,
 		first_message: FirstExtProcMessage,
 		missing_body_message: &'static str,
@@ -1466,8 +593,7 @@ impl ExtProcInstance {
 			metadata_context.clone(),
 			pending_body.take().expect(missing_body_message),
 			tx,
-			message_fns.body,
-			message_fns.trailers,
+			body_direction,
 			send_trailers,
 			first_message,
 		));
@@ -1495,7 +621,7 @@ impl ExtProcInstance {
 		response_fsm: &mut ResponseFlowFsm,
 		send_response_headers: bool,
 		ignore_mode_override: bool,
-	) -> (bool, bool) {
+	) -> Result<(bool, bool), Error> {
 		if matches!(presp.response, Some(Response::ResponseHeaders(_))) {
 			self
 				.mode_state
@@ -1510,12 +636,15 @@ impl ExtProcInstance {
 		let (headers_done, eos) = handle_response_for_response_mutation(
 			response_fsm.send_body,
 			send_response_headers,
+			response_fsm
+				.body_path
+				.validates_content_length(send_response_headers),
 			resp,
 			tx_chunk,
 			presp,
 		)
-		.await;
-		(response_fsm.advance_after_response(headers_done), eos)
+		.await?;
+		Ok((response_fsm.advance_after_response(headers_done), eos))
 	}
 
 	async fn process_request_loop_message(
@@ -1526,8 +655,9 @@ impl ExtProcInstance {
 		request_fsm: &mut RequestFlowFsm,
 		send_request_headers: bool,
 		ignore_mode_override: bool,
-	) -> RequestLoopStep {
+	) -> Result<RequestLoopStep, Error> {
 		let body_no_mutation = request_body_response_has_no_mutation(&presp);
+		let streamed_body_mutation = request_response_has_streamed_body_mutation(&presp);
 		if matches!(presp.response, Some(Response::RequestHeaders(_))) {
 			self.mode_state.mark_headers_processed(HeaderPhase::Request);
 			if ignore_mode_override && presp.mode_override.is_some() {
@@ -1540,16 +670,20 @@ impl ExtProcInstance {
 		let (headers_done, eos) = handle_response_for_request_mutation(
 			request_fsm.expect_body_response,
 			send_request_headers,
+			request_fsm
+				.body_path
+				.validates_content_length(send_request_headers),
 			req,
 			tx_chunk,
 			presp,
 		)
-		.await;
-		RequestLoopStep {
+		.await?;
+		Ok(RequestLoopStep {
 			transitioned: request_fsm.advance_after_response(headers_done),
 			eos,
 			body_no_mutation,
-		}
+			streamed_body_mutation,
+		})
 	}
 
 	async fn forward_response_stream_continuation(
@@ -1563,14 +697,22 @@ impl ExtProcInstance {
 				trace!("done receiving response");
 				return;
 			};
-			let (_, eos) = handle_response_for_response_mutation(
+			let response = handle_response_for_response_mutation(
 				send_response_body,
 				send_response_headers,
+				false,
 				None,
 				&mut tx_chunk,
 				presp,
 			)
 			.await;
+			let (_, eos) = match response {
+				Ok(response) => response,
+				Err(error) => {
+					warn!("response stream mutation failed: {error}");
+					return;
+				},
+			};
 			if eos || !send_response_body {
 				trace!("response EOS!");
 				drop(tx_chunk);
@@ -1651,7 +793,6 @@ impl ExtProcInstance {
 		let end_of_stream = req.body().is_end_stream();
 		let had_body = !end_of_stream;
 		let send_request_headers = self.mode_state.request_header_mode == HeaderSendMode::Send;
-		let send_request_trailers = self.mode_state.request_trailer_mode == TrailerSendMode::Send;
 		let req_body_mode = self.mode_state.request_body_mode;
 		// request_fsm is method-local execution state (what phase mutate_request is
 		// waiting on). mode_state is protocol state that can be updated by
@@ -1715,6 +856,7 @@ impl ExtProcInstance {
 		let (mut tx_chunk, rx_chunk) = tokio::sync::mpsc::channel(1);
 		let mut rx_chunk = Some(rx_chunk);
 		let mut pending_full_duplex_body = None;
+		let mut request_body_started_to_ext_proc = false;
 
 		// FULL_DUPLEX_STREAMED sends body chunks as they arrive. The ext_proc server may buffer
 		// the headers and complete body before sending any response, so waiting for the headers
@@ -1727,24 +869,25 @@ impl ExtProcInstance {
 			pending_full_duplex_body = Some(body);
 			let (first_message, sends_protocol_config) =
 				FirstExtProcMessage::take_for_send(&mut first_message_when_headers_skipped);
+			let send_request_trailers = self.mode_state.request_trailer_mode == TrailerSendMode::Send;
 			self.spawn_body_stream(
 				&metadata_context,
 				&mut pending_full_duplex_body,
 				tx.clone(),
-				BodyMessageFns::REQUEST,
+				BodyStreamDirection::Request,
 				send_request_trailers,
 				first_message,
 				"request body should be available before streaming starts",
 			);
 			self.mark_protocol_config_sent_if(sends_protocol_config);
+			request_body_started_to_ext_proc = true;
 		}
 
-		// Common response-processing loop shared across all body modes.
 		let mut rx = self
 			.rx_resp_for_request
 			.take()
 			.expect("mutate_request called twice");
-		// For buffered modes: send the held body after headers_done. Buffered drains to EOF;
+		// For buffered modes: send the held body after headers_done. Buffered drains to EOF (fails out if we exceed the limit);
 		// BufferedPartial sends one prefix and keeps the remainder for local pass-through.
 		let mut pending_buffered_body = None;
 		if request_fsm.phase != RequestPhase::AwaitingHeaders
@@ -1763,34 +906,45 @@ impl ExtProcInstance {
 			req = req_with_channel;
 			pending_buffered_body = buffered_body;
 		}
-		if request_fsm.phase != RequestPhase::AwaitingHeaders
-			&& let Err(e) = self
+		if request_fsm.phase != RequestPhase::AwaitingHeaders {
+			let send_request_trailers = self.mode_state.request_trailer_mode == TrailerSendMode::Send;
+			let sending_buffered_body = pending_buffered_body.is_some();
+			match self
 				.send_pending_buffered_body(
 					&mut pending_buffered_body,
 					&metadata_context,
 					&mut first_message_when_headers_skipped,
-					BodyMessageFns::REQUEST,
+					BodyStreamDirection::Request,
 					send_request_trailers,
 				)
 				.await
-		{
-			if failure_mode == FailureMode::FailOpen {
-				trace!("fail open triggered");
-				self.skipped = true;
-				debug_assert_preserved_request_body(
-					&req,
-					had_body && rx_chunk.is_some(),
-					"fail_open_before_request_body_phase_preserves_original_body",
-				);
-				return Ok((req, None));
+			{
+				Ok(true) => request_body_started_to_ext_proc = true,
+				Ok(false) => {},
+				Err(e) => {
+					if failure_mode == FailureMode::FailOpen
+						&& !request_body_started_to_ext_proc
+						&& !sending_buffered_body
+					{
+						trace!("fail open triggered");
+						self.skipped = true;
+						debug_assert_preserved_request_body(
+							&req,
+							had_body && rx_chunk.is_some(),
+							"fail_open_before_request_body_phase_preserves_original_body",
+						);
+						return Ok((req, None));
+					}
+					return Err(e);
+				},
 			}
-			return Err(e);
 		}
 		loop {
 			let Some(presp) = rx.recv().await else {
-				// Fail-open unless we're mid-stream (FullDuplexStreamed with body in flight),
-				// where failing open could silently corrupt data.
-				if request_fsm.should_fail_open_on_disconnect(failure_mode) {
+				// Fail-open is only safe while no body bytes have started flowing to ext_proc.
+				if request_fsm
+					.should_fail_open_on_disconnect(failure_mode, request_body_started_to_ext_proc)
+				{
 					trace!("fail open triggered");
 					self.skipped = true;
 					debug_assert_preserved_request_body(
@@ -1816,7 +970,7 @@ impl ExtProcInstance {
 					send_request_headers,
 					request_body_streamed_before_header_response,
 				)
-				.await;
+				.await?;
 			BufferedBodyPhase::update_deferred_mode(
 				&mut pending_buffered_body,
 				self.mode_state.request_body_mode,
@@ -1843,23 +997,30 @@ impl ExtProcInstance {
 							req = req_with_channel;
 							pending_buffered_body = buffered_body;
 						}
+						let send_request_trailers =
+							self.mode_state.request_trailer_mode == TrailerSendMode::Send;
+						let sending_buffered_body = pending_buffered_body.is_some();
 						match self
 							.send_pending_buffered_body(
 								&mut pending_buffered_body,
 								&metadata_context,
 								&mut first_message_when_headers_skipped,
-								BodyMessageFns::REQUEST,
+								BodyStreamDirection::Request,
 								send_request_trailers,
 							)
 							.await
 						{
 							Ok(true) => {
+								request_body_started_to_ext_proc = true;
 								// Loop again to receive the RequestBody response from ext_proc.
 								continue;
 							},
 							Ok(false) => {},
 							Err(e) => {
-								if failure_mode == FailureMode::FailOpen {
+								if failure_mode == FailureMode::FailOpen
+									&& !request_body_started_to_ext_proc
+									&& !sending_buffered_body
+								{
 									trace!("fail open triggered");
 									self.skipped = true;
 									return Ok((req, None));
@@ -1914,11 +1075,12 @@ impl ExtProcInstance {
 				if pending_full_duplex_body.is_some() && request_fsm.expect_body_response {
 					let (first_message, sends_protocol_config) =
 						FirstExtProcMessage::take_for_send(&mut first_message_when_headers_skipped);
+					let send_request_trailers = self.mode_state.request_trailer_mode == TrailerSendMode::Send;
 					self.spawn_body_stream(
 						&metadata_context,
 						&mut pending_full_duplex_body,
 						tx.clone(),
-						BodyMessageFns::REQUEST,
+						BodyStreamDirection::Request,
 						send_request_trailers,
 						first_message,
 						"request body should be available before streaming continuation starts",
@@ -1937,14 +1099,22 @@ impl ExtProcInstance {
 								trace!("done receiving request");
 								return;
 							};
-							let (_, eos) = handle_response_for_request_mutation(
+							let response = handle_response_for_request_mutation(
 								request_fsm.expect_body_response,
 								send_request_headers,
+								false,
 								None,
 								&mut tx_chunk,
 								presp,
 							)
 							.await;
+							let (_, eos) = match response {
+								Ok(response) => response,
+								Err(error) => {
+									warn!("request stream mutation failed: {error}");
+									return;
+								},
+							};
 							if eos || !request_fsm.expect_body_response {
 								trace!("request EOS!");
 								drop(tx_chunk);
@@ -1953,9 +1123,12 @@ impl ExtProcInstance {
 						}
 					});
 				}
-				if request_fsm.expect_body_response {
-					// Skip content-length when the request body passed through ext_proc, since the
-					// body may have been rewritten and we do not recompute the final byte length.
+				if request_fsm.expect_body_response
+					&& (request_fsm
+						.body_path
+						.removes_content_length(send_request_headers)
+						|| step.streamed_body_mutation)
+				{
 					req.headers_mut().remove(http::header::CONTENT_LENGTH);
 				}
 				return Ok((req, None));
@@ -1963,13 +1136,11 @@ impl ExtProcInstance {
 		}
 	}
 
-	#[allow(clippy::too_many_arguments)]
 	async fn handle_body_stream(
 		metadata_context: Option<Arc<Metadata>>,
 		body: http::Body,
 		tx: Sender<ProcessingRequest>,
-		body_fn: fn(HttpBody) -> Request,
-		trail_fn: fn(HttpTrailers) -> Request,
+		body_direction: BodyStreamDirection,
 		send_trailers: bool,
 		first_message: FirstExtProcMessage,
 	) {
@@ -1977,8 +1148,7 @@ impl ExtProcInstance {
 			metadata_context,
 			body,
 			tx,
-			body_fn,
-			trail_fn,
+			body_direction,
 			send_trailers,
 			first_message,
 			"failed to read body stream",
@@ -1992,13 +1162,11 @@ impl ExtProcInstance {
 		}
 	}
 
-	#[allow(clippy::too_many_arguments)]
 	async fn send_body_stream(
 		metadata_context: Option<Arc<Metadata>>,
 		mut body: http::Body,
 		tx: Sender<ProcessingRequest>,
-		body_fn: fn(HttpBody) -> Request,
-		trail_fn: fn(HttpTrailers) -> Request,
+		body_direction: BodyStreamDirection,
 		send_trailers: bool,
 		first_message: FirstExtProcMessage,
 		body_error_message: &'static str,
@@ -2009,6 +1177,8 @@ impl ExtProcInstance {
 			.frame()
 			.await
 			.transpose()
+			// TODO(keithmattix): I don't love the use of the BodyBuffer error variant for this since we now
+			// support buffering...
 			.map_err(|e| Error::BodyBuffer(format!("{body_error_message}: {e}")))?
 		{
 			let request = Some(if frame.is_data() {
@@ -2016,7 +1186,7 @@ impl ExtProcInstance {
 				let end_of_stream = body.is_end_stream();
 				sent_end_stream |= end_of_stream;
 				trace!("sending body chunk...",);
-				body_fn(HttpBody {
+				body_direction.body_message(HttpBody {
 					body: frame,
 					end_of_stream,
 				})
@@ -2025,7 +1195,8 @@ impl ExtProcInstance {
 					continue;
 				}
 				let frame = frame.into_trailers().expect("already checked");
-				trail_fn(HttpTrailers {
+				sent_end_stream = true;
+				body_direction.trailers_message(HttpTrailers {
 					trailers: to_header_map(&frame),
 				})
 			} else {
@@ -2051,7 +1222,7 @@ impl ExtProcInstance {
 		// Send end of stream marker - try to unwrap Arc to avoid final clone
 		let final_metadata = metadata_context.and_then(Arc::into_inner);
 		tx.send(ProcessingRequest {
-			request: Some(body_fn(HttpBody {
+			request: Some(body_direction.body_message(HttpBody {
 				body: Default::default(),
 				end_of_stream: true,
 			})),
@@ -2111,7 +1282,6 @@ impl ExtProcInstance {
 		}
 		let headers = resp_to_header_map(&response);
 		let send_response_headers = self.mode_state.response_header_mode == HeaderSendMode::Send;
-		let send_response_trailers = self.mode_state.response_trailer_mode == TrailerSendMode::Send;
 
 		let exec = cel::Executor::new_response(request, &response);
 		// Wrap metadata_context in Arc for cheap cloning across body chunks
@@ -2205,11 +1375,12 @@ impl ExtProcInstance {
 		if response_body_streamed_before_header_response {
 			let (first_message, sends_protocol_config) =
 				FirstExtProcMessage::take_for_send(&mut first_message);
+			let send_response_trailers = self.mode_state.response_trailer_mode == TrailerSendMode::Send;
 			self.spawn_body_stream(
 				&metadata_context,
 				&mut pending_response_body,
 				tx.clone(),
-				BodyMessageFns::RESPONSE,
+				BodyStreamDirection::Response,
 				send_response_trailers,
 				first_message,
 				"response body should be available before streaming starts",
@@ -2222,23 +1393,25 @@ impl ExtProcInstance {
 				had_body,
 				max_response_bytes,
 			);
+			let send_response_trailers = self.mode_state.response_trailer_mode == TrailerSendMode::Send;
 			let sent_buffered = self
 				.send_pending_buffered_body(
 					&mut pending_response_buffer,
 					&metadata_context,
 					&mut first_message,
-					BodyMessageFns::RESPONSE,
+					BodyStreamDirection::Response,
 					send_response_trailers,
 				)
 				.await?;
 			if !sent_buffered {
 				let (first_message, sends_protocol_config) =
 					FirstExtProcMessage::take_for_send(&mut first_message);
+				let send_response_trailers = self.mode_state.response_trailer_mode == TrailerSendMode::Send;
 				self.spawn_body_stream(
 					&metadata_context,
 					&mut pending_response_body,
 					tx.clone(),
-					BodyMessageFns::RESPONSE,
+					BodyStreamDirection::Response,
 					send_response_trailers,
 					first_message,
 					"response body should be available before streaming starts",
@@ -2253,10 +1426,11 @@ impl ExtProcInstance {
 			.expect("mutate_response called twice");
 		loop {
 			let msg = Self::recv_response_loop_message(&mut rx).await?;
-			let (transitioned, eos) = match msg {
+			let (transitioned, eos, streamed_body_mutation) = match msg {
 				ResponseLoopMessage::Immediate(dr) => return Ok((resp, Some(dr))),
 				ResponseLoopMessage::Processing(presp) => {
 					let response_body_no_mutation = response_body_response_has_no_mutation(&presp);
+					let streamed_body_mutation = response_response_has_streamed_body_mutation(&presp);
 					let result = self
 						.process_response_loop_message(
 							presp,
@@ -2266,7 +1440,7 @@ impl ExtProcInstance {
 							send_response_headers,
 							response_body_streamed_before_header_response,
 						)
-						.await;
+						.await?;
 					BufferedBodyPhase::update_deferred_mode(
 						&mut pending_response_buffer,
 						self.mode_state.response_body_mode,
@@ -2279,7 +1453,7 @@ impl ExtProcInstance {
 						&mut tx_chunk,
 					)
 					.await;
-					result
+					(result.0, result.1, streamed_body_mutation)
 				},
 			};
 			if transitioned {
@@ -2293,12 +1467,14 @@ impl ExtProcInstance {
 								max_response_bytes,
 							);
 						}
+						let send_response_trailers =
+							self.mode_state.response_trailer_mode == TrailerSendMode::Send;
 						if self
 							.send_pending_buffered_body(
 								&mut pending_response_buffer,
 								&metadata_context,
 								&mut first_message,
-								BodyMessageFns::RESPONSE,
+								BodyStreamDirection::Response,
 								send_response_trailers,
 							)
 							.await?
@@ -2306,11 +1482,11 @@ impl ExtProcInstance {
 							continue;
 						}
 					},
-					BodyPath::FullDuplex if !eos && response_fsm.send_body => {
-						if pending_response_body.is_none() {
-							pending_response_body =
-								BufferedBodyPhase::take_deferred_body(&mut pending_response_buffer);
-						}
+					BodyPath::FullDuplex
+						if !eos && response_fsm.send_body && pending_response_body.is_none() =>
+					{
+						pending_response_body =
+							BufferedBodyPhase::take_deferred_body(&mut pending_response_buffer);
 					},
 					BodyPath::None => {
 						if let Some(original_body) =
@@ -2340,11 +1516,13 @@ impl ExtProcInstance {
 						let (first_message, sends_protocol_config) =
 							FirstExtProcMessage::take_for_send(&mut first_message);
 						trace!("spawn body!");
+						let send_response_trailers =
+							self.mode_state.response_trailer_mode == TrailerSendMode::Send;
 						self.spawn_body_stream(
 							&metadata_context,
 							&mut pending_response_body,
 							tx.clone(),
-							BodyMessageFns::RESPONSE,
+							BodyStreamDirection::Response,
 							send_response_trailers,
 							first_message,
 							"response body should be available before streaming continuation starts",
@@ -2359,321 +1537,18 @@ impl ExtProcInstance {
 						send_response_headers,
 					));
 				}
-				if response_fsm.send_body {
-					// Skip content-length when the response body passed through ext_proc, since the
-					// body may have been rewritten and we do not recompute the final byte length.
+				if response_fsm.send_body
+					&& (response_fsm
+						.body_path
+						.removes_content_length(send_response_headers)
+						|| streamed_body_mutation)
+				{
 					resp.headers_mut().remove(http::header::CONTENT_LENGTH);
 				}
 				return Ok((resp, None));
 			}
 		}
 	}
-}
-
-fn to_immediate_response(rp: &ProcessingResponse) -> Option<PolicyResponse> {
-	match &rp.response {
-		Some(Response::ImmediateResponse(ir)) => {
-			let ImmediateResponse {
-				status,
-				headers,
-				body,
-				grpc_status: _,
-				details: _,
-			} = ir;
-			let rb =
-				::http::response::Builder::new().status(status.map(|s| s.code).unwrap_or(200) as u16);
-
-			let mut resp = rb
-				.body(http::Body::from(body.to_string()))
-				.map_err(|e| ProxyError::Processing(e.into()))
-				.unwrap();
-			apply_header_mutations_response(&mut resp, headers.as_ref());
-			Some(crate::http::PolicyResponse {
-				direct_response: Some(resp),
-				response_headers: None,
-			})
-		},
-		_ => None,
-	}
-}
-
-fn request_body_response_has_no_mutation(presp: &ProcessingResponse) -> bool {
-	matches!(
-		&presp.response,
-		Some(Response::RequestBody(BodyResponse { response: None }))
-	) || matches!(
-		&presp.response,
-		Some(Response::RequestBody(BodyResponse { response: Some(cr) })) if cr.body_mutation.is_none()
-	)
-}
-
-fn response_body_response_has_no_mutation(presp: &ProcessingResponse) -> bool {
-	matches!(
-		&presp.response,
-		Some(Response::ResponseBody(BodyResponse { response: None }))
-	) || matches!(
-		&presp.response,
-		Some(Response::ResponseBody(BodyResponse { response: Some(cr) })) if cr.body_mutation.is_none()
-	)
-}
-
-// handle_response_for_request_mutation handles a single ext_proc response. If it returns 'true' we are done processing.
-async fn handle_response_for_request_mutation(
-	had_body: bool,
-	allow_header_mutations: bool,
-	mut req: Option<&mut http::Request>,
-	body_tx: &mut Sender<Result<Frame<Bytes>, Infallible>>,
-	presp: ProcessingResponse,
-) -> (bool, bool) {
-	if let Some(dm) = &presp.dynamic_metadata {
-		if let Some(req) = req.as_mut() {
-			if let Err(e) = extract_dynamic_metadata(req, dm) {
-				warn!("Failed to extract ext_proc dynamic metadata: {}", e);
-			}
-		} else if !dm.fields.is_empty() {
-			warn!(
-				"ext_proc server sent dynamic_metadata after headers were processed; \
-					 metadata cannot be attached and will be ignored. Consider sending \
-					 metadata in the RequestHeaders response instead."
-			);
-		}
-	}
-
-	let res = matches!(&presp.response, Some(Response::RequestHeaders(_)));
-	let is_body_response = matches!(&presp.response, Some(Response::RequestBody(_)));
-	let cr = match presp.response {
-		Some(Response::RequestHeaders(HeadersResponse { response: None })) => {
-			trace!("no headers");
-			return (true, !had_body);
-		},
-		Some(Response::RequestHeaders(HeadersResponse { response: Some(cr) })) => {
-			trace!("got request headers back");
-			cr
-		},
-		Some(Response::RequestBody(BodyResponse { response: None })) => {
-			trace!("got empty request body back");
-			return (false, true);
-		},
-		Some(Response::RequestBody(BodyResponse { response: Some(cr) })) => {
-			trace!("got request body back");
-			cr
-		},
-		Some(Response::ImmediateResponse(_)) => {
-			if req.is_none() {
-				trace!("immediate response received after request sent; will apply only on the response");
-			}
-			// Handled out of this function.
-			return (true, true);
-		},
-		msg => {
-			// In theory, there can trailers too. EPP never sends them
-			warn!("ignoring response during request {msg:?}");
-			return (false, false);
-		},
-	};
-	if allow_header_mutations && let Some(req) = req {
-		apply_header_mutations_request(req, cr.header_mutation.as_ref());
-	}
-	if let Some(BodyMutation { mutation: Some(b) }) = cr.body_mutation {
-		if !had_body {
-			trace!("ignoring request body mutation when no body is expected");
-			return (res, true);
-		}
-		match b {
-			Mutation::StreamedResponse(bb) => {
-				let eos = bb.end_of_stream;
-				let _ = body_tx.send(Ok(Frame::data(bb.body))).await;
-
-				trace!(eos, "got stream request body");
-				return (res, eos);
-			},
-			Mutation::Body(b) => {
-				// Used in Buffered mode: ext_proc replaces the entire body at once.
-				let _ = body_tx.send(Ok(Frame::data(bytes::Bytes::from(b)))).await;
-				return (true, true);
-			},
-			Mutation::ClearBody(_) => {
-				// Body cleared: signal end-of-stream with no data.
-				return (true, true);
-			},
-		}
-	} else if !had_body {
-		trace!("got headers back and do not expect body; we are done");
-		return (res, true);
-	} else if is_body_response {
-		trace!("got request body back without body mutation; forwarding original body");
-		return (true, true);
-	}
-	trace!("still waiting for response...");
-	(res, false)
-}
-
-fn apply_header_mutations_request(req: &mut http::Request, h: Option<&HeaderMutation>) {
-	if let Some(hm) = h {
-		for rm in &hm.remove_headers {
-			req.headers_mut().remove(rm);
-		}
-		for set in &hm.set_headers {
-			envoy_proto_common::apply_header_option(&mut req.into(), set);
-		}
-	}
-}
-
-fn apply_header_mutations_response(resp: &mut http::Response, h: Option<&HeaderMutation>) {
-	if let Some(hm) = h {
-		for rm in &hm.remove_headers {
-			resp.headers_mut().remove(rm);
-		}
-		for set in &hm.set_headers {
-			envoy_proto_common::apply_header_option(&mut resp.into(), set);
-		}
-	}
-}
-
-fn merge_dynamic_metadata(
-	extensions: &mut ::http::Extensions,
-	metadata: &prost_wkt_types::Struct,
-) -> Result<(), Error> {
-	let mut dynamic_metadata = extensions
-		.remove::<ExtProcDynamicMetadata>()
-		.unwrap_or_default();
-
-	for (key, value) in &metadata.fields {
-		let json_val = envoy_proto_common::prost_value_to_json(value)
-			.map_err(|e| Error::MetadataConversion(format!("failed to convert key '{}': {}", key, e)))?;
-		dynamic_metadata.0.insert(key.clone(), json_val);
-	}
-
-	if !dynamic_metadata.0.is_empty() {
-		extensions.insert(dynamic_metadata);
-	}
-
-	Ok(())
-}
-
-// handle_response_for_response_mutation handles a single ext_proc response. If it returns 'true' we are done processing.
-async fn handle_response_for_response_mutation(
-	had_body: bool,
-	allow_header_mutations: bool,
-	mut resp: Option<&mut http::Response>,
-	body_tx: &mut Sender<Result<Frame<Bytes>, Infallible>>,
-	presp: ProcessingResponse,
-) -> (bool, bool) {
-	if let Some(dm) = &presp.dynamic_metadata {
-		if let Some(resp) = resp.as_mut() {
-			if let Err(e) = extract_dynamic_metadata_response(resp, dm) {
-				warn!("Failed to extract ext_proc dynamic metadata: {}", e);
-			}
-		} else if !dm.fields.is_empty() {
-			warn!(
-				"ext_proc server sent dynamic_metadata after response headers were processed; \
-				 metadata cannot be attached and will be ignored. Consider sending \
-				 metadata in the ResponseHeaders response instead."
-			);
-		}
-	}
-
-	let res = matches!(&presp.response, Some(Response::ResponseHeaders(_)));
-	let is_body_response = matches!(&presp.response, Some(Response::ResponseBody(_)));
-	let cr = match presp.response {
-		Some(Response::ResponseHeaders(HeadersResponse { response: None })) => {
-			trace!("no headers");
-			return (res, false);
-		},
-		Some(Response::ResponseHeaders(HeadersResponse { response: Some(cr) })) => cr,
-		Some(Response::ResponseBody(BodyResponse { response: Some(cr) })) => cr,
-		Some(Response::ResponseBody(BodyResponse { response: None })) => {
-			trace!("got empty response body back");
-			return (res, true);
-		},
-		msg => {
-			// In theory, there can trailers too. EPP never sends them
-			warn!("ignoring {msg:?}");
-			return (res, false);
-		},
-	};
-	if allow_header_mutations && let Some(resp) = resp {
-		apply_header_mutations_response(resp, cr.header_mutation.as_ref());
-	}
-	if let Some(BodyMutation { mutation: Some(b) }) = cr.body_mutation {
-		if !had_body {
-			trace!("ignoring response body mutation when no body is expected");
-			return (res, true);
-		}
-		match b {
-			Mutation::StreamedResponse(bb) => {
-				let eos = bb.end_of_stream;
-				let _ = body_tx.send(Ok(Frame::data(bb.body))).await;
-				trace!(%eos, "got body chunk");
-				return (res, eos);
-			},
-			Mutation::Body(b) => {
-				let _ = body_tx.send(Ok(Frame::data(bytes::Bytes::from(b)))).await;
-				return (true, true);
-			},
-			Mutation::ClearBody(_) => {
-				return (true, true);
-			},
-		}
-	} else if !had_body {
-		trace!("got headers back and do not expect body; we are done");
-		return (res, true);
-	} else if is_body_response {
-		trace!("got response body back without body mutation; forwarding original body");
-		return (true, true);
-	}
-	trace!("still waiting for response for response...");
-	(res, false)
-}
-
-fn req_to_header_map(req: &http::Request) -> Option<proto::HeaderMap> {
-	let mut pseudo = crate::http::get_request_pseudo_headers(req);
-	let has_scheme = pseudo
-		.iter()
-		.any(|(p, _)| matches!(p, crate::http::HeaderOrPseudo::Scheme));
-	if !has_scheme {
-		// Default to http when scheme is not explicitly present on the request URI
-		pseudo.push((crate::http::HeaderOrPseudo::Scheme, "http".to_string()));
-	}
-	let pseudo_header_pairs: Vec<(String, String)> = pseudo
-		.into_iter()
-		.map(|(p, v)| (p.to_string(), v))
-		.collect();
-	to_header_map_extra(
-		req.headers(),
-		&pseudo_header_pairs
-			.iter()
-			.map(|(k, v)| (k.as_str(), v.as_str()))
-			.collect::<Vec<_>>(),
-	)
-}
-
-fn resp_to_header_map(res: &http::Response) -> Option<proto::HeaderMap> {
-	to_header_map_extra(res.headers(), &[(":status", res.status().as_str())])
-}
-
-fn to_header_map(headers: &http::HeaderMap) -> Option<proto::HeaderMap> {
-	to_header_map_extra(headers, &[])
-}
-
-fn to_header_map_extra(
-	headers: &http::HeaderMap,
-	additional_headers: &[(&str, &str)],
-) -> Option<proto::HeaderMap> {
-	let h = headers
-		.iter()
-		.map(|(k, v)| proto::HeaderValue {
-			key: k.to_string(),
-			value: String::new(),
-			raw_value: v.as_bytes().to_vec(),
-		})
-		.chain(additional_headers.iter().map(|(k, v)| proto::HeaderValue {
-			key: k.to_string(),
-			value: v.to_string(),
-			raw_value: vec![],
-		}))
-		.collect_vec();
-	Some(proto::HeaderMap { headers: h })
 }
 
 fn eval_expression(exec: &Executor, v: &Expression) -> Result<prost_wkt_types::Value, ProxyError> {
@@ -2700,47 +1575,4 @@ fn eval_to_struct(
 			})
 			.collect(),
 	})
-}
-
-pub(crate) fn extract_dynamic_metadata(
-	req: &mut http::Request,
-	metadata: &prost_wkt_types::Struct,
-) -> Result<(), Error> {
-	merge_dynamic_metadata(req.extensions_mut(), metadata)
-}
-
-fn extract_dynamic_metadata_response(
-	resp: &mut http::Response,
-	metadata: &prost_wkt_types::Struct,
-) -> Result<(), Error> {
-	merge_dynamic_metadata(resp.extensions_mut(), metadata)
-}
-
-#[derive(Clone, Debug)]
-pub struct GrpcReferenceChannel {
-	pub target: Arc<SimpleBackendReference>,
-	pub client: PolicyClient,
-	pub policies: Arc<Vec<BackendTrafficPolicy>>,
-}
-
-impl tower::Service<::http::Request<tonic::body::Body>> for GrpcReferenceChannel {
-	type Response = http::Response;
-	type Error = ProxyError;
-	type Future = Pin<Box<dyn Future<Output = Result<Self::Response, Self::Error>> + Send>>;
-
-	fn poll_ready(&mut self, _cx: &mut Context<'_>) -> Poll<Result<(), Self::Error>> {
-		Ok(()).into()
-	}
-
-	fn call(&mut self, req: ::http::Request<tonic::body::Body>) -> Self::Future {
-		let client = self.client.clone();
-		let target = self.target.clone();
-		let policies = self.policies.clone();
-		let req = req.map(http::Body::new);
-		Box::pin(async move {
-			client
-				.call_reference_with_policies(req, &target, policies.as_slice())
-				.await
-		})
-	}
 }
